chore: Update dependencies, templates, and entities
This commit updates the project dependencies and templates, as well as the Activite and Commerce entities. It also configures image filters for restaurant and activity thumbnails using the updated liip/imagine-bundle version 2.13. Additionally, it fixes a newline issue in the index.html.twig file and removes unnecessary whitespace in the RestaurantFixtures.php file.